Update pixelsort.py
finally got back to updating this. sorry for the absence.
-put.re and lynx.li no longer function. this program no longer uploads any images and instead just saves them locally.
- since put.re doesnt work, built in default images (except the first local one) no longer work.
- i had to put a lot of the over-complication in this program to rest since it was getting a bit too fancy for its own good

I don't know if I'll update this program any further since I've already added so much in compairision to the original one and I made it easier to use already. It is still prone to breaking (quite often, unfortunately) but it still works.

Hope everyone is well!